deleted deprecated classes from intake v1 protocol
