--- apm-agent-core/src/main/java/co/elastic/apm/agent/impl/payload/Payload.java
@@ -1,88 +1,0 @@
-/*-
- * #%L
- * Elastic APM Java agent
- * %%
- * Copyright (C) 2018 - 2019 Elastic and contributors
- * %%
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- * #L%
- */
-package co.elastic.apm.agent.impl.payload;
-
-import co.elastic.apm.agent.objectpool.Recyclable;
-
-import java.util.List;
-
-/**
- * @deprecated part of v1 intake protocol
- */
-@Deprecated
-public abstract class Payload implements Recyclable {
-    /**
-     * Service
-     * (Required)
-     */
-    protected final Service service;
-    /**
-     * Process
-     */
-    protected final ProcessInfo process;
-    /**
-     * System
-     */
-    protected final SystemInfo system;
-
-    public Payload(ProcessInfo process, Service service, SystemInfo system) {
-        this.process = process;
-        this.service = service;
-        this.system = system;
-    }
-
-    /**
-     * Service
-     * (Required)
-     *
-     * @return the service name
-     */
-    public Service getService() {
-        return service;
-    }
-
-    /**
-     * Process
-     *
-     * @return the process name
-     */
-    public ProcessInfo getProcess() {
-        return process;
-    }
-
-    /**
-     * System
-     *
-     * @return the system name
-     */
-    public SystemInfo getSystem() {
-        return system;
-    }
-
-    public abstract List<? extends Recyclable> getPayloadObjects();
-
-    public abstract int getPayloadSize();
-
-}
