+++ integration-tests/application-server-integration-tests/src/test/java/co/elastic/apm/servlet/MockServerContainer.java
@@ -30,8 +30,6 @@ import org.testcontainers.containers.GenericContainer;
 
 /**
  * Acts as a mock for the APM-server.
- * It stores the {@link co.elastic.apm.agent.impl.payload.Payload}s sent to it,
- * which can be retrieved via the {@link #client}.
  */
 public class MockServerContainer extends GenericContainer<MockServerContainer> {
